Reorganized auto
