+++ src/org/team2168/PID/trajectory/QuinticTrajectory.java
@@ -82,17 +82,17 @@ public class QuinticTrajectory
 		
 //		//Square Path
 		double[][] waypointPath = new double[][]{
-			{10, 18, Math.PI/2}, //For Right switch from center 
-			{10, 19, Math.PI/2},
-			{5.5, 24, -Math.PI/4}
+			{11, 12, Math.PI/2}, //For Right switch from center 
+			{11, 20, Math.PI/2},
+			{15, 21, -Math.PI/4}
 
 		};
 		
-		double[][] waypointPath2 = new double[][]{
-			{6, 26, 2.36},
-			{5, 28, 1.79},
-			{5, 32, Math.PI/2}
-	};
+//		double[][] waypointPath2 = new double[][]{
+//			{6, 26, 2.36},
+//			{5, 28, 1.79},
+//			{5, 32, Math.PI/2}
+//	};
 		
 //		//Square Path
 //		double[][] waypointPath = new double[][]{
@@ -125,8 +125,8 @@ public class QuinticTrajectory
 		QuinticTrajectory quinticPath= new QuinticTrajectory(waypointPath);
 		quinticPath.calculate();
 
-		QuinticTrajectory quinticPath2= new QuinticTrajectory(waypointPath2);
-		quinticPath2.calculate();
+//		QuinticTrajectory quinticPath2= new QuinticTrajectory(waypointPath2);
+//		quinticPath2.calculate();
 
 		
 		
@@ -146,9 +146,10 @@ public class QuinticTrajectory
 		fig3.setYTic(0, fieldWidth, 1);
 		fig3.addData(quinticPath.rightPath, Color.magenta);
 		fig3.addData(quinticPath.leftPath, Color.blue);
-		fig3.addData(quinticPath2.leftPath, Color.blue);
-		fig3.addData(quinticPath2.rightPath, Color.magenta);
-		fig3.addData(waypointPath2, null, Color.black);
+		//fig3.addData(quinticPath2.leftPath, Color.blue);
+		//fig3.addData(quinticPath2.rightPath, Color.magenta);
+		//
+		//fig3.addData(waypointPath2, null, Color.black);
 		fig3.addData(new double[][]{{4.667, 3}}, Color.black);
