+++ src/org/team2168/Robot.java
@@ -5,7 +5,9 @@ import org.team2168.PID.trajectory.OneDimensionalMotionProfiling;
 import org.team2168.PID.trajectory.QuinticTrajectory;
 import org.team2168.commands.auto.*;
 import org.team2168.commands.auto.massComp.DriveStraight;
+import org.team2168.commands.auto.massComp.DriveToLeftSwitch;
 import org.team2168.commands.auto.massComp.DriveToLeftSwitchAndRightScaleFromLeft;
+import org.team2168.commands.auto.massComp.DriveToRightSwitch;
 import org.team2168.commands.auto.selector.TestAutoCommandGroupA;
 import org.team2168.commands.auto.selector.AutoStartCenter1Cube;
 import org.team2168.commands.auto.selector.AutoStartLeft2Cube;
