+++ src/org/team2168/commands/auto/selector/AutoStartLeft1Cube.java
@@ -1,8 +1,8 @@
 package org.team2168.commands.auto.selector;
 
 import org.team2168.Robot;
-import org.team2168.commands.auto.DriveToLeftSwitch;
-import org.team2168.commands.auto.DriveToRightSwitch;
+import org.team2168.commands.auto.massComp.DriveToLeftSwitch;
+import org.team2168.commands.auto.massComp.DriveToRightSwitch;
 import org.team2168.commands.auto.massComp.LeftScaleOnlyFromLeftSide;
 import org.team2168.commands.auto.massComp.RightScaleOnlyFromLeftSide;
